move eventDetails page from /admin to /member
